feat: periods 1/2 (#10)
* start periods

* fixes

* getting there

* missing revalidations

* adjust formatting rules

* fix

* more fix

+++ src/app/(overview)/_actions.ts
@@ -1,12 +1,13 @@
 "use server";
 
 import { revalidateTag } from "next/cache";
-import { eq } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { parseAsync } from "valibot";
 import type { Input } from "valibot";
 
 import { db } from "~/db";
-import { timeslot } from "~/db/schema";
+import { getOpenPeriods } from "~/db/getters";
+import { period, timeslot } from "~/db/schema";
 import { currentUser } from "~/lib/auth";
 import type { CurrencyCode } from "~/lib/currencies";
 import { currencies } from "~/lib/currencies";
@@ -24,6 +25,20 @@ export async function reportTime(props: Input<typeof reportTimeSchema>) {
   const normalizedAmount =
     input.chargeRate * currency.base ** currency.exponent;
 
+  const slotPeriod = await db.query.period.findFirst({
+    where: and(
+      eq(period.status, "open"),
+      eq(period.tenantId, user.id),
+      eq(period.clientId, input.clientId),
+    ),
+  });
+  if (!slotPeriod) {
+    // TODO: Create a new one
+    throw new Error("No open period found");
+  }
+
+  console.log("Inserting timeslot for period", slotPeriod.id);
+
   await db.insert(timeslot).values({
     date: input.date,
     duration: String(input.duration),
@@ -32,22 +47,30 @@ export async function reportTime(props: Input<typeof reportTimeSchema>) {
     description: input.description,
     clientId: input.clientId,
     tenantId: user.id,
+    periodId: slotPeriod.id,
   });
 
   revalidateTag("timeslots");
-  console.log("[server]: returning from action");
+  revalidateTag("periods");
 }
 
 export async function deleteTimeslot(id: number) {
+  const user = await currentUser();
+  if (!user) return;
+
   await db.delete(timeslot).where(eq(timeslot.id, id));
 
   revalidateTag("timeslots");
+  revalidateTag("periods");
 }
 
 export async function updateTimeslot(
   id: number,
   props: Input<typeof updateSchema>,
 ) {
+  const user = await currentUser();
+  if (!user) return;
+
   const input = await parseAsync(updateSchema, props);
   const currency = input.currency
     ? currencies[input.currency as CurrencyCode]
@@ -67,4 +90,33 @@ export async function updateTimeslot(
     .where(eq(timeslot.id, id));
 
   revalidateTag("timeslots");
+  revalidateTag("periods");
+}
+
+export async function closePeriod(
+  id: number,
+  props: {
+    openNewPeriod: boolean;
+  },
+) {
+  const user = await currentUser();
+  if (!user) return;
+
+  const openPeriods = await getOpenPeriods(user.id);
+  const p = openPeriods.find((p) => p.id === id);
+
+  if (!p) {
+    throw new Error("Period not found");
+  }
+
+  await db
+    .update(period)
+    .set({ status: "closed", closedAt: new Date() })
+    .where(eq(period.id, id));
+
+  if (props.openNewPeriod) {
+    // TODO:
+  }
+
+  revalidateTag("periods");
 }
